refactor(api): move /api/prescription/prescribers/{userId} to deprecated context
