+++ api/src/deprecated/infrastructure/serializers/jsonapi/prescriber-serializer.js
