+++ api/src/team/application/routes.js
@@ -8,7 +8,6 @@ import { membershipRoutes } from './membership/membership.route.js';
 import { organizationMemberIdentitiesRoute } from './membership/organization-member-identities.route.js';
 import { organizationInvitationAdminRoutes } from './organization-invitations/organization-invitation.admin.route.js';
 import { organizationInvitationRoutes } from './organization-invitations/organization-invitation.route.js';
-import { prescriberInformationsRoute } from './prescriber-informations.route.js';
 import { userOrgaSettingsRoute } from './user-orga-settings.route.js';
 
 const register = async function (server) {
@@ -22,7 +21,6 @@ const register = async function (server) {
     ...membershipRoutes,
     ...organizationMemberIdentitiesRoute,
     ...userOrgaSettingsRoute,
-    ...prescriberInformationsRoute,
     ...organizationInvitationRoutes,
     ...organizationInvitationAdminRoutes,
   ]);
