+++ api/tests/deprecated/unit/application/prescriber-informations.controller.test.js
@@ -1,29 +1,21 @@
 import sinon from 'sinon';
 
-import { prescriberInformationsController } from '../../../../src/team/application/prescriber-informations.controller.js';
-import { usecases } from '../../../../src/team/domain/usecases/index.js';
+import { prescriberInformationsController } from '../../../../src/deprecated/application/prescriber-informations.controller.js';
+import { usecases } from '../../../../src/deprecated/domain/usecases/index.js';
 import { expect } from '../../../test-helper.js';
 import { hFake } from '../../../tooling/mocks/hapi.mock.js';
 
-describe('Unit | Team | Application | Controller | prescriber-informations', function () {
+describe('Deprecated | Unit | Application | Controller | prescriber-informations', function () {
   describe('#get', function () {
-    let request;
-
-    beforeEach(function () {
-      request = { auth: { credentials: { userId: 1 } } };
-
-      sinon.stub(usecases, 'getPrescriber');
-    });
-
     it('should get the prescriber', async function () {
       // given
-      const prescriberSerializer = {
-        serialize: sinon.stub(),
-      };
+      sinon.stub(usecases, 'getPrescriber');
       usecases.getPrescriber.withArgs({ userId: 1 }).resolves({});
+      const prescriberSerializer = { serialize: sinon.stub() };
       prescriberSerializer.serialize.withArgs({}).returns('ok');
 
       // when
+      const request = { auth: { credentials: { userId: 1 } } };
       const response = await prescriberInformationsController.get(request, hFake, { prescriberSerializer });
 
       // then
