+++ api/tests/deprecated/unit/infrastructure/serializers/jsonapi/prescriber-serializer.test.js
@@ -1,12 +1,12 @@
+import { prescriberSerializer } from '../../../../../../src/deprecated/infrastructure/serializers/jsonapi/prescriber-serializer.js';
 import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../../../src/identity-access-management/domain/constants/identity-providers.js';
 import { Organization } from '../../../../../../src/organizational-entities/domain/models/Organization.js';
 import { ORGANIZATION_FEATURE } from '../../../../../../src/shared/constants.js';
 import { Membership } from '../../../../../../src/shared/domain/models/Membership.js';
-import { prescriberSerializer } from '../../../../../../src/team/infrastructure/serializers/jsonapi/prescriber-serializer.js';
 import { expect } from '../../../../../test-helper.js';
 import { domainBuilder } from '../../../../../tooling/domain-builder/domain-builder.js';
 
-describe('Unit | Team | Infrastructure | Serializer | JSONAPI | prescriber', function () {
+describe('Deprecated | Unit | Infrastructure | Serializer | JSONAPI | prescriber', function () {
   describe('#serialize', function () {
     let user;
